fix(tests): api_coverage all-params row asserts the documented f32 narrowing of temperature — a round trip of 0.7 is f64::from(0.7f32) by contract
Train run 34658698055: workspace-test (intel) test_completion_request_with_all_params: left Some(0.699999988079071) right Some(0.7). deserialize_temperature_f64 narrows through f32 on purpose (types.rs: 'the value still narrows to f32 before it reaches a sampler, so the narrowing is checked here too') and returns the narrowed value, so 0.7 can never round-trip exactly; the row was dark (#3130 class) and could never pass. It now asserts the contract. Local: api_coverage 270/270.

Pmat-Ticket: PMAT-1098

+++ crates/aprender-serve/tests/api_coverage.rs
@@ -3779,7 +3779,12 @@ fn test_completion_request_with_all_params() {
     let deserialized: CompletionRequest = serde_json::from_str(&json).expect("should deserialize");
 
     assert_eq!(deserialized.max_tokens, Some(256));
-    assert_eq!(deserialized.temperature, Some(0.7));
+    // The temperature deserializer narrows through f32 on purpose (types.rs
+    // deserialize_temperature_f64: "the value still narrows to f32 before it reaches a
+    // sampler, so the narrowing is checked here too"), so a round trip of 0.7 yields
+    // f64::from(0.7f32), never 0.7 exactly. Asserting the narrowed value is asserting
+    // the contract; asserting 0.7 was a dark row that could never pass (#3130, train #3127).
+    assert_eq!(deserialized.temperature, Some(f64::from(0.7_f32)));
 }
 
 #[test]
